Deploy Latest (#942)
Deploys swap-erc20 across mainnets;
Improves deploy script usability;
Adds arbitrum to constants and configs;
Uses key rather than mnemonic for hardhat;
Fixes MAX_ERROR_COUNT in swap-erc20;
Removes ERC721 from wrapper;
Removes ALCHEMY_API_KEY;
Removes contributors in favor of code owners;
Removes remaining rinkeby metadata;

+++ source/pool/scripts/deploy.js
@@ -1,8 +1,10 @@
 /* eslint-disable no-console */
 const fs = require('fs')
+const Confirm = require('prompt-confirm')
 const { ethers, run } = require('hardhat')
 const stakingDeploys = require('@airswap/staking/deploys.js')
 const { chainNames, stakingTokenAddresses } = require('@airswap/constants')
+const { getEtherscanURL } = require('@airswap/utils')
 const poolDeploys = require('../deploys.js')
 
 async function main() {
@@ -11,35 +13,46 @@ async function main() {
   console.log(`Deployer: ${deployer.address}`)
 
   const chainId = await deployer.getChainId()
+  const gasPrice = await deployer.getGasPrice()
   const scale = 10
   const max = 100
   const stakingContract = stakingDeploys[chainId]
   const stakingToken = stakingTokenAddresses[chainId]
 
   console.log(`Deploying on ${chainNames[chainId].toUpperCase()}`)
+  console.log(`Staking token: ${stakingToken}`)
+  console.log(`Staking contract: ${stakingContract}`)
+  console.log(`Gas price: ${gasPrice / 10 ** 9} gwei`)
 
-  const poolFactory = await ethers.getContractFactory('Pool')
-  const poolContract = await poolFactory.deploy(
-    scale,
-    max,
-    stakingContract,
-    stakingToken
-  )
-  await poolContract.deployed()
-  console.log(`Deployed: ${poolContract.address}`)
+  const prompt = new Confirm('Proceed to deploy?')
+  if (await prompt.run()) {
+    const poolFactory = await ethers.getContractFactory('Pool')
+    const poolContract = await poolFactory.deploy(
+      scale,
+      max,
+      stakingContract,
+      stakingToken
+    )
+    console.log(
+      'Deploying...',
+      getEtherscanURL(chainId, poolContract.deployTransaction.hash)
+    )
+    await poolContract.deployed()
+    console.log(`Deployed: ${poolContract.address}`)
 
-  poolDeploys[chainId] = poolContract.address
-  fs.writeFileSync(
-    './deploys.js',
-    `module.exports = ${JSON.stringify(poolDeploys, null, '\t')}`
-  )
-  console.log('Updated deploys.js')
+    poolDeploys[chainId] = poolContract.address
+    fs.writeFileSync(
+      './deploys.js',
+      `module.exports = ${JSON.stringify(poolDeploys, null, '\t')}`
+    )
+    console.log('Updated deploys.js')
 
-  console.log(
-    `\nVerify with "yarn verify --network ${chainNames[
-      chainId
-    ].toLowerCase()}"\n`
-  )
+    console.log(
+      `\nVerify with "yarn verify --network ${chainNames[
+        chainId
+      ].toLowerCase()}"\n`
+    )
+  }
 }
 
 main()
