+++ source/swap-erc20/scripts/deploy.js
@@ -1,9 +1,11 @@
 /* eslint-disable no-console */
 const fs = require('fs')
+const Confirm = require('prompt-confirm')
 const { ethers, run } = require('hardhat')
 const poolDeploys = require('@airswap/pool/deploys.js')
 const stakingDeploys = require('@airswap/staking/deploys.js')
 const { chainNames } = require('@airswap/constants')
+const { getEtherscanURL } = require('@airswap/utils')
 const swapDeploys = require('../deploys.js')
 
 async function main() {
@@ -12,6 +14,7 @@ async function main() {
   console.log(`Deployer: ${deployer.address}`)
 
   const chainId = await deployer.getChainId()
+  const gasPrice = await deployer.getGasPrice()
   const protocolFeeWallet = poolDeploys[chainId]
   const stakingContract = stakingDeploys[chainId]
   const protocolFee = 7
@@ -20,33 +23,41 @@ async function main() {
   const rebateMax = 100
 
   console.log(`Deploying on ${chainNames[chainId].toUpperCase()}`)
-  console.log(`Fee wallet (Pool): ${protocolFeeWallet}`)
-  console.log(`Staking: ${stakingContract}`)
+  console.log(`Fee recipient: ${protocolFeeWallet}`)
+  console.log(`Staking contract: ${stakingContract}`)
+  console.log(`Gas price: ${gasPrice / 10 ** 9} gwei`)
 
-  const swapFactory = await ethers.getContractFactory('Swap')
-  const swapContract = await swapFactory.deploy(
-    protocolFee,
-    protocolFeeLight,
-    protocolFeeWallet,
-    rebateScale,
-    rebateMax,
-    stakingContract
-  )
-  await swapContract.deployed()
-  console.log(`Deployed: ${swapContract.address}`)
+  const prompt = new Confirm('Proceed to deploy?')
+  if (await prompt.run()) {
+    const swapFactory = await ethers.getContractFactory('SwapERC20')
+    const swapContract = await swapFactory.deploy(
+      protocolFee,
+      protocolFeeLight,
+      protocolFeeWallet,
+      rebateScale,
+      rebateMax,
+      stakingContract
+    )
+    console.log(
+      'Deploying...',
+      getEtherscanURL(chainId, swapContract.deployTransaction.hash)
+    )
+    await swapContract.deployed()
+    console.log(`Deployed: ${swapContract.address}`)
 
-  swapDeploys[chainId] = swapContract.address
-  fs.writeFileSync(
-    './deploys.js',
-    `module.exports = ${JSON.stringify(swapDeploys, null, '\t')}`
-  )
-  console.log('Updated deploys.js')
+    swapDeploys[chainId] = swapContract.address
+    fs.writeFileSync(
+      './deploys.js',
+      `module.exports = ${JSON.stringify(swapDeploys, null, '\t')}`
+    )
+    console.log('Updated deploys.js')
 
-  console.log(
-    `\nVerify with "yarn verify --network ${chainNames[
-      chainId
-    ].toLowerCase()}"\n`
-  )
+    console.log(
+      `\nVerify with "yarn verify --network ${chainNames[
+        chainId
+      ].toLowerCase()}"\n`
+    )
+  }
 }
 
 main()
